rename AmplifyProvider folder to ThemeProvider, update tests
