+++ packages/react/src/components/index.tsx
@@ -1,3 +1,3 @@
-export * from './AmplifyProvider';
+export * from './ThemeProvider';
 export * from './Authenticator';
 export * from './Geo';
